Move parametric distributions into a distributions subpackage
Pure restructuring, no logic changes:

- Move the 16 distribution modules (weibull, gamma, normal, etc. and
  custom_distribution) from surpyval/univariate/parametric/ into
  surpyval/univariate/parametric/distributions/
- Re-export all names from the new subpackage so the public API,
  including surpyval.univariate.parametric.*, is unchanged
- Convert the relative imports in bernoulli and exact_event_time to
  parent-relative imports
- Update Sphinx autoclass paths in docs/univariate to the new module
  locations

The core model classes (parametric.py, parametric_fitter.py), the
fitters subpackage, and the model combinators (mixture_model, series,
parallel) remain at the package top level.
