+++ docs/univariate/expo_weibull.rst
@@ -1,3 +1,3 @@
-.. autoclass:: surpyval.univariate.parametric.expo_weibull.ExpoWeibull_
+.. autoclass:: surpyval.univariate.parametric.distributions.expo_weibull.ExpoWeibull_
    :members:
    :inherited-members:
