+++ docs/univariate/gamma.rst
@@ -1,3 +1,3 @@
-.. autoclass:: surpyval.univariate.parametric.gamma.Gamma_
+.. autoclass:: surpyval.univariate.parametric.distributions.gamma.Gamma_
    :members:
    :inherited-members:
