+++ docs/univariate/gumbel.rst
@@ -1,3 +1,3 @@
-.. autoclass:: surpyval.univariate.parametric.gumbel.Gumbel_
+.. autoclass:: surpyval.univariate.parametric.distributions.gumbel.Gumbel_
    :members:
    :inherited-members:
