+++ docs/univariate/logistic.rst
@@ -1,3 +1,3 @@
-.. autoclass:: surpyval.univariate.parametric.logistic.Logistic_
+.. autoclass:: surpyval.univariate.parametric.distributions.logistic.Logistic_
    :members:
    :inherited-members:
