+++ surpyval/univariate/parametric/__init__.py
@@ -13,27 +13,30 @@ Parametric Analysis
 
 import numpy as np
 
-from .bernoulli import Bernoulli, FixedEventProbability
-from .beta import Beta
-from .custom_distribution import CustomDistribution
-from .exact_event_time import ExactEventTime
-from .expo_weibull import ExpoWeibull
-from .exponential import Exponential
-from .gamma import Gamma
-from .gumbel import Gumbel
-from .gumbel_lev import GumbelLEV
-from .logistic import Logistic
-from .loglogistic import LogLogistic
-from .lognormal import Galton, LogNormal
+from .distributions import (
+    Bernoulli,
+    Beta,
+    CustomDistribution,
+    ExactEventTime,
+    ExpoWeibull,
+    Exponential,
+    FixedEventProbability,
+    Galton,
+    Gamma,
+    Gauss,
+    Gumbel,
+    GumbelLEV,
+    Logistic,
+    LogLogistic,
+    LogNormal,
+    Normal,
+    Rayleigh,
+    Uniform,
+    Weibull,
+)
 from .mixture_model import MixtureModel
-from .normal import Gauss, Normal
 from .parametric import Parametric
 from .parametric_fitter import ParametricFitter
-from .rayleigh import Rayleigh
-from .uniform import Uniform
-from .weibull import Weibull
-
-# from surpyval.univariate import nonparametric as nonp
 
 
 class NeverOccurs:
