+++ surpyval/univariate/parametric/distributions/logistic.py
