+++ surpyval/univariate/parametric/distributions/loglogistic.py
