+++ surpyval/univariate/parametric/distributions/rayleigh.py
